Run Jekyll build from Gulp, instead of server
Spawn a child process to run Jekyll build during the 'build' Gulp task.
This removes the need to run two servers during dev (BrowserSync
proxying to the Jekyll server), which was causing a race-condition
during the rebuild/reload cycle. We can also simplify the npm scripts to
simply call the Gulp tasks.