contact info acts like a sticky footer
Contact links and photo will follow after the about content, but on
larger screens will stick to the bottom of the viewport. The about
content will expand to take the remaining top portion of the screen,
with its content vertically centred in that space. Padding is applied to
the top of the about section to ensure a minimum distance from the top
of the page.