Conditionally apply flexbox rules to contact links
Use @supports to only apply flexbox rules to browsers that support it.
This check is not strictly required, as these rules will gracefully fallback
when not supported. However, the transform rules depend on the items being
positioned with flexbox, so we only want to apply those in supported
browsers. It made sense to group these together.